requirements.txt
torch
torchvision
opencv-python
pandas
numpy
tqdm
pillow
matplotlib
